requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
firebase-admin==6.4.0
langchain==0.1.0
langchain-google-genai==0.0.6
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0